robo/requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
grpcio==1.16.1
grpcio-tools==1.16.1
idna==2.8
protobuf==3.9.2
pycparser==2.19
pymongo==3.9.0
pyOpenSSL==19.0.0
PySocks==1.7.1
requests==2.22.0
six==1.12.0
urllib3==1.24.2
